# backend/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "DicDoc Service",
        "title": "DicDoc Service",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/v1/appointments": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получить все записи из таблицы Appointments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "appointments"
                ],
                "summary": "Получить записи",
                "operationId": "get-appointments",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор врача",
                        "name": "doctor_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Идентификатор пациента",
                        "name": "patient_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.GetAppointmentResponseDto"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создать запись пациента к врачу",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "appointments"
                ],
                "summary": "Создать запись",
                "operationId": "create-appointment",
                "parameters": [
                    {
                        "description": "Информация о создаваемой записи",
                        "name": "Appointment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateOrEditAppointmentRequestDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateOrEditAppointmentResponseDto"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    }
                }
            }
        },
        "/api/v1/appointments/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получить запись по идентификатору в таблице Appointments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "appointments"
                ],
                "summary": "Получение записи по идентификатору",
                "operationId": "get-appointments-by-id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор записи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Запись",
                        "schema": {
                            "$ref": "#/definitions/dto.GetAppointmentResponseDto"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "404": {
                        "description": "Запись не найдена",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Изменить запись в таблице Appointments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "appointments"
                ],
                "summary": "Изменить запись",
                "operationId": "edit-appointment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор записи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Информация о изменяемой записи",
                        "name": "Appointment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateOrEditAppointmentRequestDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateOrEditAppointmentResponseDto"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "403": {
                        "description": "Пользователь не имеет доступа",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "404": {
                        "description": "Запись не найдена",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удалить запись в таблице Appointments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "appointments"
                ],
                "summary": "Удалить запись",
                "operationId": "cancel-appointment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор записи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Запись успешно удалена",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "404": {
                        "description": "Запись не найдена",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.CreateOrEditAppointmentRequestDto": {
            "description": "Информация о записи при создании или изменении",
            "type": "object",
            "properties": {
                "dateTime": {
                    "description": "Дата и время записи",
                    "type": "string"
                },
                "doctorId": {
                    "description": "Идентификатор врача",
                    "type": "integer"
                },
                "id": {
                    "description": "Идентификатор записи",
                    "type": "integer"
                },
                "patientId": {
                    "description": "Идентификатор пациента",
                    "type": "integer"
                }
            }
        },
        "dto.CreateOrEditAppointmentResponseDto": {
            "description": "Информация о записи при создании или изменении",
            "type": "object",
            "properties": {
                "id": {
                    "description": "Идентификатор записи",
                    "type": "integer"
                }
            }
        },
...
        "dto.ErrorDto": {
            "description": "Информация об ошибке (DTO)",
            "type": "object",
            "properties": {
                "error": {
                    "description": "Ошибка",
                    "type": "string"
                }
            }
        },
        "dto.GetAppointmentResponseDto": {
            "description": "Информация о записи при получении",
            "type": "object",
            "properties": {
                "dateTime": {
                    "description": "Дата и время записи",
                    "type": "string"
                },
                "doctorId": {
                    "description": "Идентификатор врача",
                    "type": "integer"
                },
                "id": {
                    "description": "Идентификатор записи",
                    "type": "integer"
                },
                "patientId": {
                    "description": "Идентификатор пациента",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Authorization check",
            "type": "apiKey",
            "name": "auth-x",
            "in": "header"
        }
    }
}